fix(tasks): update info passed to the tasks comments notification context
